send_allで使っている変数の型を見直し

+++ echo_server/utils.h
@@ -1,7 +1,7 @@
 #ifndef UTILS_H
 #define UTILS_H
 
-int send_all(int socket_fd, char *buffer, int length);
+int send_all(int socket_fd, char *buffer, size_t length);
 int recv_all(int socket_fd, char *buffer, int length);
 int valid_port(const char *port_str, int *port);
 int create_socket();
